r10460/sofa : Fix a bunch of "loss of data" warnings in framework/
Here starts the great journey of fixing the bazillion warnings due
to not using size_t. I took no risk and fixed only simple cases.


Former-commit-id: e74770d303293b688cee5f5392301103f5264b7b

+++ framework/sofa/core/objectmodel/AspectPool.h
@@ -106,7 +106,7 @@ public:
 
     AspectRef getAspect(int id);
 
-    int nbAspects() const { return aspects.size(); }
+    int nbAspects() const { return (int) aspects.size(); }
     int getAspectCounter(int id) const { return aspects[id]->counter; }
 
     friend class Aspect;
